HEALIX_HMS/healix-hms/services/auth_service/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
bcrypt
